requirements.txt
requests
azure-mgmt-network
azure-identity
python-dotenv